requirements_prod.txt
fastapi==0.128.0
uvicorn==0.40.0
python-multipart==0.0.22
torch --index-url https://download.pytorch.org/whl/cpu
torchvision --index-url https://download.pytorch.org/whl/cpu
pytorch-lightning==2.6.0
opencv-python-headless==4.12.0.88
albumentations==2.0.8
pandas==2.3.3
numpy==2.2.6
pillow==12.1.0